requirements.txt
ccxt
openai
python-dotenv
schedule
backtrader
numpy
pandas
ta
accelerate